thoughts_app_service/requirements.txt
adrf==0.1.5
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-property==0.2.2
certifi==2024.2.2
colored==2.2.4
colorlog==6.8.2
distro==1.9.0
Django==5.0.3
django-prometheus==2.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
inflection==0.5.1
openai==1.14.3
packaging==24.0
prometheus_client==0.20.0
psycopg2==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
uritemplate==4.1.1
boto3==1.34.72